Release 0.1.1: notebooks, Binder wiring, float exponent ergonomics
- Add notebooks/{01_getting_started,02_historical_prices,03_websocket_streaming}.ipynb
  mirroring the polymarket-pandas style. Binder badge now launches directly
  into 01_getting_started.ipynb.
- Coerce `exponent` to Float64 in /latest_price, /price, /reduce_price so
  `10 ** df["exponent"]` works without a cast.
- Bump to 0.1.1; update CHANGELOG with compare links.

+++ examples/quickstart.py
@@ -22,10 +22,7 @@ def main() -> None:
             properties=["price", "confidence", "exponent", "publisherCount"],
             formats=[],
         )
-        # Convert raw mantissa to a human-readable price.
-        # Cast to float — exponents are negative ints, and `int ** negative-int`
-        # raises in numpy/pandas. Float exponentiation handles it correctly.
-        df["humanPrice"] = df["price"].astype(float) * 10.0 ** df["exponent"].astype(float)
+        df["humanPrice"] = df["price"] * 10 ** df["exponent"]
         print(df[["priceFeedId", "humanPrice", "publisherCount"]])
         print("update timestamp (µs):", df.attrs.get("timestampUs"))
